# Auth-Login/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Auth Login Protect API",
    "version": "1.0.0",
    "description": "Secure API with Supabase Auth, JWT verification, protected routes, and Swagger UI."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Server is running"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Sign up a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "test@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "test@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid login credentials"
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public info",
        "responses": {
          "200": {
            "description": "Public endpoint response"
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get protected profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Protected profile returned"
          },
          "401": {
            "description": "Missing, invalid, or expired token"
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get protected dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Protected dashboard returned"
          },
          "401": {
            "description": "Missing, invalid, or expired token"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out the current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Missing, invalid, or expired token"
          }
        }
      }
    }
  }
}